# src/spiffworkflow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: Workflow Microservice
  license:
    name: MIT
servers:
  - url: http://localhost:5000/v1.0
security:
  - jwt: ["secret"]
paths:
...
  /process-groups:
    parameters:
      - name: page
        in: query
        required: false
        description: The page number to return. Defaults to page 1.
        schema:
          type: integer
      - name: per_page
        in: query
        required: false
        description: The number of groups to show per page. Defaults to page 10.
        schema:
          type: integer
    # process_groups_list
    get:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_groups_list
      summary: get list
      tags:
        - Process Groups
      responses:
        "200":
          description: An array of process groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowSpecCategory"
    # process_group_add
    post:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_group_add
      summary: Add process group
      tags:
        - Process Groups
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowSpecCategory"
      responses:
        "201":
          description: Workflow spec category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSpecCategory"
  /process-groups/{process_group_id}:
    parameters:
      - name: process_group_id
        in: path
        required: true
        description: The unique id of an existing process group.
        schema:
          type: string
    # process_group_show
    get:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_group_show
      summary: Returns a single process group
      tags:
        - Process Groups
      responses:
        "200":
          description: Workflow spec category.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSpecCategory"
    delete:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_group_delete
      summary: Deletes a single process group
      tags:
        - Process Groups
      responses:
        "204":
          description: The process group was deleted.
    put:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_group_update
      summary: Upates a single process group
      tags:
        - Process Groups
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowSpecCategory"
      responses:
        "200":
          description: Process group updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSpecCategory"
  /process-groups/{process_group_id}/process-models:
    parameters:
      - name: process_group_id
        in: path
        required: true
        description: The group containing the models we want to return
        schema:
          type: string
      - name: page
        in: query
        required: false
        description: The page number to return. Defaults to page 1.
        schema:
          type: integer
      - name: per_page
        in: query
        required: false
        description: The number of models to show per page. Defaults to page 10.
        schema:
          type: integer
    # process_model_list
    get:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_model_list
      summary: Return a list of process models for a given process group
      tags:
        - Process Models
      responses:
        "200":
          description: Successfully return the requested models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowSpec"
  /process-models:
    # process_model_add
    post:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_model_add
      summary: Creates a new workflow specification with the given parameters.
      tags:
        - Process Models
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowSpec"
      responses:
        "201":
          description: Workflow specification created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSpec"
  /process-models/{process_group_id}/{process_model_id}/file:
    parameters:
      - name: process_group_id
        in: path
        required: true
        description: The group containing the models we want to return
        schema:
          type: string
      - name: process_model_id
        in: path
        required: true
        description: The unique id of an existing workflow specification to validate.
        schema:
          type: string
    # add_file
    post:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.add_file
      summary: Add a new workflow spec file
      tags:
        - Spec Files
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Metadata about the uploaded file, but not the file content.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/File"
    # get:
    #   operationId: spiffworkflow_backend.api.process_api_blueprint.get_files
    #   summary: Provide a list of workflow spec files for the given workflow_spec_id. IMPORTANT, only includes metadata, not the file content.
    #   tags:
    #     - Spec Files
    #   responses:
    #     '200':
    #       description: An array of file descriptions (not the file content)
    #       content:
    #         application/json:
    #          schema:
    #            type: array
    #            items:
    #              $ref: "#/components/schemas/File"
  /process-models/{process_group_id}/{process_model_id}:
    parameters:
      - name: process_group_id
        in: path
        required: true
        description: The unique id of an existing process group
        schema:
          type: string
      - name: process_model_id
        in: path
        required: true
        description: The unique id of an existing process model.
        schema:
          type: string
    # process_model_show
    get:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_model_show
      summary: Returns a single process model
      tags:
        - Process Models
      responses:
        "200":
          description: Workflow spec.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSpec"
    # process_instance_create
    post:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_instance_create
      summary: Creates an process instance from a process model and returns the instance
      tags:
        - Process Instances
      responses:
        "201":
          description: Workflow generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    # process_model_delete
    delete:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_model_delete
      summary: Removes an existing process model
      tags:
        - Process Models
      responses:
        "200":
          description: The workflow specification has been removed.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/OkTrue"
    # process model update
    put:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_model_update
      summary: Modifies an existing process mosel with the given parameters.
      tags:
        - Process Models
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowSpec"
      responses:
        "200":
          description: Process model updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowSpec"
  /process-models/{process_group_id}/{process_model_id}/process-instances:
    parameters:
      - name: process_group_id
        in: path
        required: true
        description: The unique id of an existing process group
        schema:
          type: string
      - name: process_model_id
        in: path
        required: true
        description: The unique id of an existing workflow specification.
        schema:
          type: string
      - name: page
        in: query
        required: false
        description: The page number to return. Defaults to page 1.
        schema:
          type: integer
      - name: per_page
        in: query
        required: false
        description: The page number to return. Defaults to page 1.
        schema:
          type: integer
      - name: start_from
        in: query
        required: false
        description: For filtering - beginning of start window - in seconds since epoch
        schema:
          type: integer
      - name: start_till
        in: query
        required: false
        description: For filtering - end of start window - in seconds since epoch
        schema:
          type: integer
      - name: end_from
        in: query
        required: false
        description: For filtering - beginning of end window - in seconds since epoch
        schema:
          type: integer
      - name: end_till
        in: query
        required: false
        description: For filtering - end of end window - in seconds since epoch
        schema:
          type: integer
      - name: process_status
        in: query
        required: false
        description: For filtering - not_started, user_input_required, waiting, complete, faulted, or suspended
        schema:
          type: string
    # process_instance_list
    get:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_instance_list
      summary: Returns a list of process instances for a given process model
      tags:
        - Process Instances
      responses:
        "200":
          description: Workflow.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
  /process-models/{process_group_id}/{process_model_id}/process-instances/{process_instance_id}:
    parameters:
      - name: process_group_id
        in: path
        required: true
        description: The unique id of an existing process group
        schema:
          type: string
      - name: process_model_id
        in: path
        required: true
        description: The unique id of an existing process model.
        schema:
          type: string
      - name: process_instance_id
        in: path
        required: true
        description: The unique id of an existing process instance.
        schema:
          type: integer
    # process_instance_delete
    delete:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_instance_delete
      summary: Deletes a single process instance
      tags:
        - Process Instances
      responses:
        "200":
          description: The process instance was deleted.
          content:
            application/json:
              schema:
                $ref: "#components/schemas/OkTrue"
  /process-models/{process_group_id}/{process_model_id}/process-instances/{process_instance_id}/run:
    parameters:
      - name: process_group_id
        in: path
        required: true
        description: The unique id of an existing process group
        schema:
          type: string
      - name: process_model_id
        in: path
        required: true
        description: The unique id of an existing workflow specification.
        schema:
          type: string
      - name: process_instance_id
        in: path
        required: true
        description: The unique id of an existing process instance.
        schema:
          type: integer
      - name: do_engine_steps
        in: query
        required: false
        description: Defaults to true, can be set to false if you are just looking at the workflow not completeing it.
        schema:
          type: boolean
    # process_instance_run
    post:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_instance_run
      summary: Run a process instance
      tags:
        - Process Instances
      responses:
        "200":
          description: Returns details about the workflows state and current task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /process-models/{process_group_id}/{process_model_id}/process-instances/report:
    parameters:
      - name: process_group_id
        in: path
        required: true
        description: The unique id of an existing process group
        schema:
          type: string
      - name: process_model_id
        in: path
        required: true
        description: The unique id of an existing workflow specification.
        schema:
          type: string
      - name: page
        in: query
        required: false
        description: The page number to return. Defaults to page 1.
        schema:
          type: integer
      - name: per_page
        in: query
        required: false
        description: The page number to return. Defaults to page 1.
        schema:
          type: integer
    get:
      operationId: spiffworkflow_backend.routes.process_api_blueprint.process_instance_report
      summary: Returns a report of process instances for a given process model
      tags:
        - Process Instances
      responses:
        "200":
          description: Workflow.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: spiffworkflow_backend.routes.user.verify_token
  schemas:
...
    WorkflowSpec:
      properties:
        id:
          type: string
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        primary_process_id:
          type: string
          nullable: true
        category_id:
          type: string
          nullable: true
        standalone:
          type: boolean
          example: false
          default: false
        workflow_spec_category:
          $ref: "#/components/schemas/WorkflowSpecCategory"
        is_status:
          type: boolean
          nullable: true
    WorkflowSpecCategory:
      properties:
        id:
          type: string
        name:
          type: string
        display_name:
          type: string
        display_order:
          type: integer
        workflows:
          type: array
          items:
            $ref: "#/components/schemas/Workflow"
    File:
      properties:
        id:
          type: number
        name:
          type: string
          example: "random_fact.bpmn"
        version:
          type: integer
        updated_at_in_seconds:
          type: string
          format: date_time
          example: "2019-12-25T09:12:33.001Z"
        type:
          type: string
        primary:
          type: boolean
        content_type:
          type: string
          example: "application/xml"
        workflow_spec_id:
          type: string
          example: "random_fact"
          x-nullable: true
        file:
          type: string
...